rename list_expression.* -> list.*
Preparation for moving list operations out of procedural.*
